Progression
Writing the work for Modello 2.0, which includes a rewrite of the engine and some demos
