+++ demo/templates/content-one.tmp
@@ -1,0 +1,3 @@
+<p>
+    Hey there! Here's some sample content.
+</p>
